requirements.txt
colorama==0.4.6
PyMySQL==1.1.1
python-dotenv==1.1.0